+++ AlphaFS/Filesystem/Path Class/Path.Combine.cs
@@ -21,7 +21,6 @@
 
 using System;
 using System.Security;
-using System.Text;
 
 namespace Alphaleonis.Win32.Filesystem
 {
@@ -36,76 +35,6 @@ namespace Alphaleonis.Win32.Filesystem
       public static string Combine(params string[] paths)
       {
          return CombineCore(true, paths);
-      }
-
-
-      /// <summary>Combines an array of strings into a path.</summary>
-      /// <returns>The combined paths.</returns>
-      /// <remarks>
-      ///   <para>The parameters are not parsed if they have white space.</para>
-      ///   <para>Therefore, if path2 includes white space (for example, " c:\\ "),</para>
-      ///   <para>the Combine method appends path2 to path1 instead of returning only path2.</para>
-      /// </remarks>
-      /// <exception cref="ArgumentNullException"/>
-      /// <exception cref="ArgumentException"/>
-      /// <param name="checkInvalidPathChars"><c>true</c> will not check <paramref name="paths"/> for invalid path characters.</param>
-      /// <param name="paths">An array of parts of the path.</param>
-      [SecurityCritical]
-      internal static string CombineCore(bool checkInvalidPathChars, params string[] paths)
-      {
-         if (paths == null)
-            throw new ArgumentNullException("paths");
-
-         var capacity = 0;
-         var num = 0;
-
-         for (int index = 0, l = paths.Length; index < l; ++index)
-         {
-            if (null == paths[index])
-               throw new ArgumentNullException("paths");
-
-            if (paths[index].Length != 0)
-            {
-               if (IsPathRooted(paths[index], checkInvalidPathChars))
-               {
-                  num = index;
-                  capacity = paths[index].Length;
-               }
-
-               else
-                  capacity += paths[index].Length;
-
-
-               var ch = paths[index][paths[index].Length - 1];
-
-               if (!IsDVsc(ch, null))
-                  ++capacity;
-            }
-         }
-
-
-         var buffer = new StringBuilder(capacity);
-
-         for (var index = num; index < paths.Length; ++index)
-         {
-            if (paths[index].Length != 0)
-            {
-               if (buffer.Length == 0)
-                  buffer.Append(paths[index]);
-
-               else
-               {
-                  var ch = buffer[buffer.Length - 1];
-
-                  if (!IsDVsc(ch, null))
-                     buffer.Append(DirectorySeparatorChar);
-
-                  buffer.Append(paths[index]);
-               }
-            }
-         }
-
-         return buffer.ToString();
       }
    }
 }